requirements.txt
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dnspython==2.7.0
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.5
fsspec==2025.5.1
ftfy==6.3.1
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mongoengine==0.29.1
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
open_clip_torch==2.32.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
timm==1.0.17
tokenizers==0.21.2
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.166
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wikipedia==1.4.0
